Extract selection probe modules and keep copy policy capability-based.
Move pure clipboard/AX policy and the hotkey selection pipeline out of TextInputIO, drop unsafe line-copy clipboard heuristics, and tighten focus fallbacks so empty-selection editors no longer invent selected text.
